💄 Feat: slider에 의한 ui 변경
slider를 사용하여 [관심있는 아이돌 추가] 부분을 api로 호출하였습니다
