+++ package.json
@@ -21,7 +21,9 @@
 		"gsap": "^3.12.7",
 		"react": "^19.0.0",
 		"react-dom": "^19.0.0",
-		"react-router-dom": "^7.5.0"
+		"react-router-dom": "^7.5.0",
+		"react-slick": "^0.30.3",
+		"slick-carousel": "^1.8.1"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^1.9.4",
